Rename importer and use __main__
